Rearrange modules by flask blueprint
