+++ setup.py
@@ -55,6 +55,7 @@ setup(name='piperider',
               'google-cloud-bigquery-storage'
           ],
           'dev': [
+              'requests'
               'pytest>=4.6',
               'pytest-flake8',
               'flake8==3.9.2',
